+++ src/CacheIdentityClient.cs
@@ -1,10 +1,12 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using System;
+using System.Threading;
+using System.Threading.Tasks;
+using Azure.Core;
 using Microsoft.Identity.Client;
 using Microsoft.Identity.Client.AppConfig;
-using System;
-using System.Threading.Tasks;
 
 namespace Microsoft.Azure.StackExchangeRedis;
 
@@ -14,18 +16,17 @@ namespace Microsoft.Azure.StackExchangeRedis;
 internal interface ICacheIdentityClient
 {
     /// <summary>
-    /// Acquire a token from the Microsoft Identity Client.
+    /// Acquire a token to be used to authenticate a connection.
     /// </summary>
-    /// <param name="forceRefresh">Pass <see langword="false"/> to quickly acquire a cached token for a new connection, and <see langword="true"/> for subsequent token refreshes.</param>
-    /// <returns>Authentication result containing a token.</returns>
-    Task<AuthenticationResult> GetTokenAsync(bool forceRefresh);
+    /// <returns>A TokenResult containing a token and expiration</returns>
+    Task<TokenResult> GetTokenAsync();
 }
 
 internal class CacheIdentityClient : ICacheIdentityClient
 {
     private static readonly string[] s_azureCacheForRedisScopes = { "acca5fbb-b7e4-4009-81f1-37e38fd66d78/.default" };
 
-    private readonly Func<bool, Task<AuthenticationResult>> _getToken;
+    private readonly Func<Task<TokenResult>> _getToken;
 
     internal static ICacheIdentityClient CreateForSystemAssignedManagedIdentity()
         => new CacheIdentityClient(ManagedIdentityApplicationBuilder.Create(ManagedIdentityId.SystemAssigned)
@@ -41,17 +42,21 @@ internal class CacheIdentityClient : ICacheIdentityClient
             .WithClientSecret(secret)
             .Build());
 
+    internal static ICacheIdentityClient CreateForTokenCredential(TokenCredential tokenCredential)
+        => new CacheIdentityClient(tokenCredential);
+
     private CacheIdentityClient(IManagedIdentityApplication managedIdentityApplication)
-        => _getToken = async (bool forceRefresh) => await managedIdentityApplication.AcquireTokenForManagedIdentity(s_azureCacheForRedisScopes[0])
-            .WithForceRefresh(forceRefresh)
-            .ExecuteAsync().ConfigureAwait(false);
+        => _getToken = async () => new TokenResult(await managedIdentityApplication.AcquireTokenForManagedIdentity(s_azureCacheForRedisScopes[0])
+            .ExecuteAsync().ConfigureAwait(false));
 
     private CacheIdentityClient(IConfidentialClientApplication confidentialClientApplication)
-        => _getToken = async (bool forceRefresh) => await confidentialClientApplication!.AcquireTokenForClient(s_azureCacheForRedisScopes)
-            .WithForceRefresh(forceRefresh)
-            .ExecuteAsync().ConfigureAwait(false);
+        => _getToken = async () => new TokenResult(await confidentialClientApplication!.AcquireTokenForClient(s_azureCacheForRedisScopes)
+            .ExecuteAsync().ConfigureAwait(false));
 
-    public async Task<AuthenticationResult> GetTokenAsync(bool forceRefresh)
-        => await _getToken.Invoke(forceRefresh).ConfigureAwait(false);
+    private CacheIdentityClient(TokenCredential tokenCredential)
+        => _getToken = async () => new TokenResult(await tokenCredential.GetTokenAsync(new TokenRequestContext(s_azureCacheForRedisScopes), CancellationToken.None));
+
+    public async Task<TokenResult> GetTokenAsync()
+        => await _getToken.Invoke().ConfigureAwait(false);
 
 }