+++ RELEASENOTES.md
@@ -4,6 +4,9 @@
 
 ## 1.1.1
 - Dynamic token refresh margin ([#3 by philon-msft](https://github.com/Azure/Microsoft.Azure.StackExchangeRedis/pull/3))
+- Remove unnecessary await in README ([#21 by eerhardt](https://github.com/Azure/Microsoft.Azure.StackExchangeRedis/pull/21))
+- Upgrade Microsoft.Identity.Client package reference ([#7 by philon-msft](https://github.com/Azure/Microsoft.Azure.StackExchangeRedis/pull/7))
+- TokenCredential support ([#10 by lsannicolas](https://github.com/Azure/Microsoft.Azure.StackExchangeRedis/pull/10))
 
 ## 1.1.0
 - AAD Support ([#1 by philon-msft](https://github.com/Azure/Microsoft.Azure.StackExchangeRedis/pull/1))